artemis/requirements.txt
attrdict
asyncpg
discord
redis
tabulate
termcolor
pyfiglet
pyyaml